requirements.txt
flask
python-dotenv
plaid-python
anthropic
gunicorn
psycopg2-binary
